Refactor code - move AutoSuggest input into separate file to be able to reuse it in update/create purposes
