+++ client/src/components/items/item/Item.js
@@ -1,3 +1,4 @@
+import { v4 as uuidv4 } from 'uuid';
 import { useDispatch } from 'react-redux';
 import { useParams } from 'react-router';
 import { useState } from 'react';
@@ -5,22 +6,41 @@ import { Link } from 'react-router-dom';
 import PropTypes from 'prop-types';
 import { updateItem, deleteItem } from '../../../redux/items/itemsSlice';
 import ClickOutside from '../../../helpers/ClickOutside';
+import TagInput from '../TagInput';
 
 const Item = ({ item }) => {
   const dispatch = useDispatch();
   const { userId, collId } = useParams();
   const [onEdit, setOnEdit] = useState(false);
-  const { _id, title, text } = item;
+  const {
+    _id, title, text, tags,
+  } = item;
   const [itemData, setItemData] = useState(item);
+  const [newTags, setNewTags] = useState(tags.map((tag) => tag.tagname));
+  const [value, setValue] = useState('');
   const handleSubmit = (e) => {
     e.preventDefault();
     dispatch(updateItem({
-      userId, collId, itemId: _id, updatedItem: itemData,
+      userId, collId, itemId: _id, updatedItem: { ...itemData, tags: newTags },
     }));
     setOnEdit(false);
   };
+  const pushTag = (tag) => {
+    setNewTags([...newTags, tag]);
+    setValue('');
+  };
+  const excludeTag = (toDelete) => {
+    setNewTags(newTags.filter((tag) => tag !== toDelete));
+  };
   const updateForm = (
-    <ClickOutside onClick={() => { setOnEdit(false); setItemData(item); }}>
+    <ClickOutside
+      onClick={() => {
+        setOnEdit(false);
+        setItemData(item);
+        setValue('');
+        setNewTags([]);
+      }}
+    >
       <form onSubmit={handleSubmit}>
         <input
           required
@@ -33,6 +53,19 @@ const Item = ({ item }) => {
           placeholder={text}
           value={itemData.text}
           onChange={(e) => setItemData({ ...itemData, text: e.target.value })}
+        />
+        <ul>
+          {newTags.map((tag) => (
+            <li key={uuidv4()}>
+              {tag}
+              <button type="button" onClick={() => excludeTag(tag)}>X</button>
+            </li>
+          ))}
+        </ul>
+        <TagInput
+          pushTag={pushTag}
+          value={value}
+          setValue={setValue}
         />
         <button type="submit">Save</button>
       </form>
@@ -48,6 +81,7 @@ const Item = ({ item }) => {
             <p>{_id}</p>
             <p>{title}</p>
             <p>{text}</p>
+            {tags.map((tag) => <span key={uuidv4()}>{tag.tagname}</span>)}
           </Link>
           <button type="button" onClick={() => setOnEdit(true)}>Edit</button>
         </div>
@@ -62,6 +96,9 @@ Item.propTypes = {
     _id: PropTypes.string.isRequired,
     title: PropTypes.string.isRequired,
     text: PropTypes.string.isRequired,
+    tags: PropTypes.arrayOf(PropTypes.shape({
+      tagname: PropTypes.string.isRequired,
+    })),
   }).isRequired,
 };
